# backend/openapi.yaml
openapi: '3.1.0'
info:
  title: SmartAccess API
  version: 1.0.0
  description: |
    IoT access-control backend providing device management, event processing,
    real-time telemetry, alerting, audit logging, and administration.
  contact:
    name: SmartAccess Team
servers:
  - url: /api
    description: Authentication endpoints (non-versioned)
  - url: /api/v1
    description: Main versioned API

# ── Reusable Components ────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived (15 min) access token
  schemas:
    ProblemDetails:
      description: RFC 7807 Problem Details
      type: object
      required: [type, title, status, detail]
      properties:
        type:
          type: string
          format: uri
          example: https://api.smartaccess.io/errors/unauthorized
        title:
          type: string
          example: Unauthorized
        status:
          type: integer
          example: 401
        detail:
          type: string
          example: Missing or invalid Authorization header
        instance:
          type: string
          example: /api/v1/devices
    AuthUser:
      type: object
      properties:
        id: { type: integer }
        email: { type: string, format: email }
        role: { type: string, enum: [ADMIN, OPERATOR, VIEWER] }
    LoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            accessToken: { type: string }
            refreshToken: { type: string }
            user: { $ref: '#/components/schemas/AuthUser' }
    Device:
      type: object
      properties:
        id: { type: integer }
        deviceUuid: { type: string, format: uuid }
        name: { type: string }
        location: { type: string, nullable: true }
        status: { type: string, enum: [REGISTERED, ONLINE, OFFLINE, ERROR, MAINTENANCE, DECOMMISSIONED] }
        firmwareVersion: { type: string, nullable: true }
        lastSeenAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetails' }
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive token pair
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Exchange refresh token for a new token pair
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: Tokens refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      accessToken: { type: string }
                      refreshToken: { type: string }
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Revoke refresh token
      operationId: logout
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        '204':
          description: Logged out

  # ── Devices ──────────────────────────────────────────────────────────────
...
  /v1/devices/{uuid}/status:
    patch:
      tags: [Devices]
      summary: Update device status
      operationId: updateDeviceStatus
      security: [{ bearerAuth: [] }]
      parameters:
        - name: uuid
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [REGISTERED, ONLINE, OFFLINE, ERROR, MAINTENANCE, DECOMMISSIONED]
      responses:
        '200':
          description: Updated device
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { $ref: '#/components/schemas/Device' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422':
          $ref: '#/components/responses/NotFound'

  # ── Events ───────────────────────────────────────────────────────────────
...